Add preload option and use relative paths in data .pt
'preload' defaults to True, which maintains the existing functionality.
If set to False then the audio data is not loaded into memory before training,
only when the that specific audio is needed for training. This slows the training.

Paths to the audio and tsv data now stored relatively so that the audo data .pt
files are more portable between runs. Thus you can move the base Maestro directory
and the generate .pt files will still be usable.